% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d8_functions.R
\name{accum}
\alias{accum}
\title{Determine flow accumulation on digital elevation models}
\usage{
accum(dirs, mode = "d8")
}
\arguments{
\item{dirs}{RasterLayer object with flow directions.}

\item{mode}{Only 'd8' supported for now.}
}
\value{
accum RasterLayer object with flow accumulation.
}
\description{
Determine flow accumulation on digital elevation models
}
