% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depressions.R
\name{breach}
\alias{breach}
\title{Remove depressions from digital elevation models by breaching depressions}
\usage{
breach(dem)
}
\arguments{
\item{dem}{RasterLayer object containing the digital elevation model.}
}
\value{
dem_breach RasterLayer object with depressions filled.
}
\description{
Remove depressions from digital elevation models by breaching depressions
}
