% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comp_breach_lindsay2016}
\alias{comp_breach_lindsay2016}
\title{Complete breaching algorithm:
"Lindsay, J.B., 2016. Efficient hybrid breaching-filling sink removal methods for flow path enforcement in digital elevation models: Efficient Hybrid Sink Removal Methods for Flow Path Enforcement. Hydrological Processes 30, 846--857. doi:10.1002/hyp.10648"
As implemented in RichDEM}
\usage{
comp_breach_lindsay2016(dem)
}
\arguments{
\item{dem}{The input digital elevation model (DEM)}
}
\value{
The DEM with depressions breached
}
\description{
Complete breaching algorithm:
"Lindsay, J.B., 2016. Efficient hybrid breaching-filling sink removal methods for flow path enforcement in digital elevation models: Efficient Hybrid Sink Removal Methods for Flow Path Enforcement. Hydrological Processes 30, 846--857. doi:10.1002/hyp.10648"
As implemented in RichDEM
}
