% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d8_flow_accum}
\alias{d8_flow_accum}
\title{Function for determining d8 flow accumulation (RichDEM)}
\usage{
d8_flow_accum(flowdirs)
}
\arguments{
\item{flowdirs}{The d8 pointer flow direction raster}
}
\value{
a flow accumulation raster
}
\description{
Function for determining d8 flow accumulation (RichDEM)
}
