% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d8_flow_directions}
\alias{d8_flow_directions}
\title{Function for determining d8 flow directions (RichDEM)}
\usage{
d8_flow_directions(dem)
}
\arguments{
\item{dem}{The input digital elevation model (DEM)}
}
\value{
a d8 flow direction raster
}
\description{
Function for determining d8 flow directions (RichDEM)
}
