% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression_functions.R
\name{dem_accumulation}
\alias{dem_accumulation}
\title{Determine flow accumulation on digital elevation models}
\usage{
dem_accumulation(raster, mode = "d8")
}
\arguments{
\item{raster}{RasterLayer object.}

\item{mode}{Only 'd8' supported for now.}
}
\value{
raster_accumulate RasterLayer object with flow accumulation.
}
\description{
Determine flow accumulation on digital elevation models
}
