% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression_functions.R
\name{dem_breach}
\alias{dem_breach}
\title{Remove depressions from digital elevation models by breaching depressions}
\usage{
dem_breach(raster)
}
\arguments{
\item{raster}{RasterLayer object.}
}
\value{
raster_fill RasterLayer object with depressions filled.
}
\description{
Remove depressions from digital elevation models by breaching depressions
}
