% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression_functions.R
\name{dem_direction}
\alias{dem_direction}
\title{Determine flow directions on digital elevation models}
\usage{
dem_direction(raster, mode = "d8")
}
\arguments{
\item{raster}{RasterLayer object.}

\item{mode}{Only 'd8' supported for now.}
}
\value{
raster_direction RasterLayer object with flow directions.
}
\description{
Determine flow directions on digital elevation models
}
