% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression_functions.R
\name{dem_fill}
\alias{dem_fill}
\title{Remove depressions from digital elevation models by filling using the Priority-Flood algorithm}
\usage{
dem_fill(raster, epsilon = TRUE)
}
\arguments{
\item{raster}{RasterLayer object.}

\item{epsilon}{TRUE (default) or FALSE. Indicates whether cell elevations in depressions should be increased to ensure drainage.}
}
\value{
raster_fill RasterLayer object with depressions filled.
}
\description{
Remove depressions from digital elevation models by filling using the Priority-Flood algorithm
}
