% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression_functions.R
\name{dem_fill_basins}
\alias{dem_fill_basins}
\title{Remove depressions from digital elevation models by filling using the Priority-Flood algorithm and get basin labels}
\usage{
dem_fill_basins(raster)
}
\arguments{
\item{raster}{RasterLayer object.}
}
\value{
stack RasterStack object two layers: one with the filled input dem and one integer raster with basin labels
}
\description{
Remove depressions from digital elevation models by filling using the Priority-Flood algorithm and get basin labels
}
