% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression_functions.R
\name{dem_fill_watersheds}
\alias{dem_fill_watersheds}
\title{Remove depressions from digital elevation models by filling using the Priority-Flood algorithm and get watershed labels}
\usage{
dem_fill_watersheds(raster)
}
\arguments{
\item{raster}{RasterLayer object.}
}
\value{
stack RasterStack object two layers: one with the filled input dem and one integer raster with watershed labels
}
\description{
Remove depressions from digital elevation models by filling using the Priority-Flood algorithm and get watershed labels
}
