% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression_functions.R
\name{dem_watershed}
\alias{dem_watershed}
\title{Delineate watersheds (same as catchment, upslope area, contributing area) from a flow direction raster and target area given as point, polygon or raster object.}
\usage{
dem_watershed(direction, target, vector_out = FALSE, nested = FALSE)
}
\arguments{
\item{direction}{RasterLayer object containing d8 flow direction.}

\item{target}{RasterLayer, or sf POINT, POLYGON or MULTIPOLYGON object.}

\item{nested}{TRUE or FALSE (default). Indicates whether the output watersheds should be nested (a watershed for each pour-point).}

\item{vector_result}{TRUE or FALSE (default). Indicates whether the output watershed should be converted to a vector geometry.}
}
\value{
raster_watershed RasterLayer object delineated watershed.
}
\description{
Delineate watersheds (same as catchment, upslope area, contributing area) from a flow direction raster and target area given as point, polygon or raster object.
}
