% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d8_functions.R
\name{dirs}
\alias{dirs}
\title{Determine flow directions on digital elevation models}
\usage{
dirs(dem, mode = "d8")
}
\arguments{
\item{dem}{RasterLayer object containing the digital elevation model.}
}
\value{
dirs RasterLayer object with flow directions.
}
\description{
Determine flow directions on digital elevation models
}
