% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depressions.R
\name{fill}
\alias{fill}
\title{Remove depressions from a digital elevation model by filling it inwards from the edges using the Priority-Flood algorithm}
\usage{
fill(dem, epsilon = TRUE)
}
\arguments{
\item{dem}{RasterLayer object containing the digital elevation model.}

\item{epsilon}{TRUE (default) or FALSE. If TRUE, cell elevations in depressions are be increased to ensure drainage. If FALSE, filled depressions are left as flat surfaces.}
}
\value{
dem_fill RasterLayer object containing the digital elevation model with depressions removed.
}
\description{
Remove depressions from a digital elevation model by filling it inwards from the edges using the Priority-Flood algorithm
}
