% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d8_functions.R
\name{dirs}
\alias{dirs}
\title{Determine flow directions}
\usage{
dirs(dem, mode = "d8")
}
\arguments{
\item{dem}{terra::SpatRaster object containing the digital elevation model.}

\item{mode}{Only 'd8' supported for now.}
}
\value{
dirs terra::SpatRaster object with flow directions.
}
\description{
Determine flow directions on digital elevation models
}
