% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depressions.R
\name{fill_basins}
\alias{fill_basins}
\title{Remove depressions by filling and delineate drainage basins}
\usage{
fill_basins(dem)
}
\arguments{
\item{dem}{terra::SpatRaster object containing the digital elevation model.}
}
\value{
dem_fill_basins terra::SpatRaster object with two layers: one with the filled input dem and one integer raster with basin labels
}
\description{
Remove depressions from a digital elevation model by filling it inwards from the edges using the Priority-Flood algorithm, and delineate drainage basins simultaneously.
}
