% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d8_functions.R
\name{watershed}
\alias{watershed}
\title{Delineate watersheds (same as catchment, upslope area, contributing area) from a flow direction raster and target area given as vector or raster object.}
\usage{
watershed(dirs, target, nested = FALSE, mode = "d8")
}
\arguments{
\item{dirs}{RasterLayer object containing d8 flow direction.}

\item{target}{RasterLayer, Spatial* or sf object.}

\item{nested}{TRUE or FALSE (default). Indicates whether the output watersheds should be nested (a watershed for each pour-point).}

\item{mode}{Only 'd8' supported for now.}
}
\value{
watershed RasterLayer object delineated watershed.
}
\description{
Delineate watersheds (same as catchment, upslope area, contributing area) from a flow direction raster and target area given as vector or raster object.
}
